requirements/common.txt
pygame-ce
pygame_gui
pygbag
